add a RDS to Parquet file format conversion script
This is to make it easier to support publishing DiAna data in the Apache
Parquet data format. The script is also added to document
exactly how the Parquet files are created.

The script converts all the RDS (`.rds`) files found in an input directory
to corresponding Parquet (`.parquet`) files in an output directory.